Better support for finding libmp4v2 library
We were checking OS type and trying random extensions (so, dll, so.0, etc). As
a result we had to have a big case statement switching on OS name, and that
was kind of brittle (and didn't work on OS X). Ctypes has a utility function
for this, so we're using that now instead.